Change public statuses pages to mount the web UI
