+++ app/views/shared/_web_app.html.haml
@@ -1,0 +1,17 @@
+- content_for :header_tags do
+  - if user_signed_in?
+    = preload_pack_asset 'features/compose.js', crossorigin: 'anonymous'
+    = preload_pack_asset 'features/home_timeline.js', crossorigin: 'anonymous'
+    = preload_pack_asset 'features/notifications.js', crossorigin: 'anonymous'
+
+  %meta{ name: 'applicationServerKey', content: Rails.configuration.x.vapid_public_key }
+
+  = render_initial_state
+  = javascript_pack_tag 'application', crossorigin: 'anonymous'
+
+.notranslate.app-holder#mastodon{ data: { props: Oj.dump(default_props) } }
+  %noscript
+    = image_pack_tag 'logo.svg', alt: 'Mastodon'
+
+    %div
+      = t('errors.noscript_html', apps_path: 'https://joinmastodon.org/apps')
